# contract/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "http",
        "description" : "Session token issued by POST /api/v1/sessions, sent as 'Authorization: Bearer <token>'.",
        "scheme" : "bearer"
      }
    },
    "schemas" : {
...
      "CursorDirectionDto" : {
        "type" : "string",
        "enum" : [ "FORWARD", "BACKWARD" ]
      },
      "CursorDto" : {
        "type" : "object",
        "required" : [ "pivotId", "direction" ],
        "properties" : {
          "pivotId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "direction" : {
            "$ref" : "#/components/schemas/CursorDirectionDto"
          }
        }
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "PaginationOutputDto" : {
        "type" : "object",
        "required" : [ "previousCursor", "nextCursor" ],
        "properties" : {
          "previousCursor" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/CursorDto"
            }, {
              "type" : "null"
            } ]
          },
          "nextCursor" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/CursorDto"
            }, {
              "type" : "null"
            } ]
          }
        }
      },
...
      "ProblemDetail" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : [ "string", "null" ]
          },
          "instance" : {
            "type" : "string"
          },
          "code" : {
            "type" : "string"
          },
          "currentLength" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          }
        },
        "required" : [ "title", "status", "detail", "instance", "code" ]
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
      "UserDataExportListOutputDto" : {
        "type" : "object",
        "required" : [ "exports", "pagination" ],
        "properties" : {
          "exports" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserDataExportOutputDto"
            }
          },
          "pagination" : {
            "$ref" : "#/components/schemas/PaginationOutputDto"
          }
        }
      },
      "UserDataExportOutputDto" : {
        "type" : "object",
        "required" : [ "id", "state", "requestedAt", "completedAt", "expiresAt", "byteSize", "mediaType", "sha256", "failureCode", "formatVersion" ],
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "state" : {
            "type" : "string"
          },
          "requestedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "completedAt" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Instant"
            }, {
              "type" : "null"
            } ]
          },
          "expiresAt" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Instant"
            }, {
              "type" : "null"
            } ]
          },
          "byteSize" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          },
          "mediaType" : {
            "type" : [ "string", "null" ]
          },
          "sha256" : {
            "type" : [ "string", "null" ]
          },
          "failureCode" : {
            "type" : [ "string", "null" ]
          },
          "formatVersion" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/me/exports" : {
      "post" : {
        "summary" : "Request an export of the caller's data",
        "description" : "Queued, not built: the archive is downloadable once the export reads READY. Needs the reauthentication header.",
        "parameters" : [ {
          "name" : "X-Reauthentication",
          "in" : "header",
          "schema" : {
            "type" : [ "string", "null" ]
          },
          "required" : true
        } ],
        "responses" : {
          "202" : {
            "description" : "Export requested and queued",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDataExportOutputDto"
                }
              }
            }
          },
          "400" : {
            "description" : "UNSUPPORTED_REAUTHENTICATION_FACTOR: the reauthentication header names no password factor",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403" : {
            "description" : "REAUTHENTICATION_FAILED: the reauthentication header is absent or its password is wrong",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409" : {
            "description" : "EXPORT_ALREADY_IN_PROGRESS: this account already has a pending export",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429" : {
            "description" : "EXPORT_TOO_SOON: the last request is inside exports.minimum_interval, and Retry-After names the wait",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "tags" : [ "Me Export Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "get" : {
        "parameters" : [ {
          "name" : "cursor",
          "in" : "query",
          "schema" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/CursorDto"
            }, {
              "type" : "null"
            } ]
          }
        }, {
          "name" : "pageSize",
          "in" : "query",
          "schema" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "One page of the caller's exports, every state included",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDataExportListOutputDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "summary" : "List Exports",
        "tags" : [ "Me Export Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/v1/me/exports/{id}" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The export's state and, once READY, its size, digest and expiry",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDataExportOutputDto"
                }
              }
            }
          },
          "403" : {
            "description" : "EXPORT_INSUFFICIENT_PERMISSIONS: this export belongs to another account",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "EXPORT_DOES_NOT_EXIST: no export of the caller carries this id",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "summary" : "Get Export",
        "tags" : [ "Me Export Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Export deleted, its archive released"
          },
          "403" : {
            "description" : "EXPORT_INSUFFICIENT_PERMISSIONS: this export belongs to another account",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "EXPORT_DOES_NOT_EXIST: no export of the caller carries this id",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "summary" : "Delete Export",
        "tags" : [ "Me Export Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
    "/api/v1/me/exports/{id}/download" : {
      "get" : {
        "summary" : "Download the archive",
        "description" : "Whole, or one byte range through the Range header. ETag carries the archive's SHA-256.",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "Range",
          "in" : "header",
          "schema" : {
            "type" : [ "string", "null" ]
          },
          "required" : true
        } ],
        "responses" : {
          "200" : {
            "description" : "The whole archive",
            "content" : {
              "application/zip" : { }
            }
          },
          "206" : {
            "description" : "The requested byte range, Content-Range set",
            "content" : {
              "application/zip" : { }
            }
          },
          "403" : {
            "description" : "EXPORT_INSUFFICIENT_PERMISSIONS: this export belongs to another account",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404" : {
            "description" : "EXPORT_DOES_NOT_EXIST: no export of the caller carries this id",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409" : {
            "description" : "EXPORT_NOT_READY: the export is still pending or has failed",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "410" : {
            "description" : "EXPORT_GONE: the archive expired, was deleted or was superseded by a newer export",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "416" : {
            "description" : "RANGE_NOT_SATISFIABLE: the Range header names bytes past the archive's end",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          }
        },
        "tags" : [ "Me Export Controller" ],
        "security" : [ {
          "SecurityScheme" : [ ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "api-application API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}